requirements.txt
Flask
flask-cors
flasgger
elevenlabs
google-generativeai